refactor: NBT-80 entity 디렉토리를 domain 디렉토리로 패키지명 변경
