+++ src/main/java/com/newbit/column/repository/ColumnRepository.java
@@ -1,6 +1,6 @@
 package com.newbit.column.repository;
 
-import com.newbit.column.entity.Column;
+import com.newbit.column.domain.Column;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ColumnRepository extends JpaRepository<Column, Long> {}
