+++ src/main/java/com/newbit/column/repository/ColumnRequestRepository.java
@@ -1,6 +1,6 @@
 package com.newbit.column.repository;
 
-import com.newbit.column.entity.ColumnRequest;
+import com.newbit.column.domain.ColumnRequest;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ColumnRequestRepository extends JpaRepository<ColumnRequest, Long> {}
