+++ src/main/java/com/newbit/column/service/ColumnRequestService.java
@@ -6,8 +6,8 @@ import com.newbit.column.dto.request.UpdateColumnRequestDto;
 import com.newbit.column.dto.response.CreateColumnResponseDto;
 import com.newbit.column.dto.response.DeleteColumnResponseDto;
 import com.newbit.column.dto.response.UpdateColumnResponseDto;
-import com.newbit.column.entity.Column;
-import com.newbit.column.entity.ColumnRequest;
+import com.newbit.column.domain.Column;
+import com.newbit.column.domain.ColumnRequest;
 import com.newbit.column.enums.RequestType;
 import com.newbit.column.mapper.ColumnMapper;
 import com.newbit.column.repository.ColumnRepository;
